Verify close() works in ws close handler
It looks like we're failing to send websocket close() events in the
webSocketClose() handler for hibernatable websockets. Print debugging
suggests we queue a close message to be sent but then never get past
waiting for the GatedMessage::outputLock promise.

Strangely, the test seems to pass if a 1ms wait is added between a
websocket send and a websocket close.